moved TInitializeOnCopy into its own header
